fix: must cancel outstanding alarm even if delay > MAX_DELAY
As the comment in the code says, delays greater than MAX_DELAY can be simply
ignored, since MAX_DELAY is so far in the future it makes no difference in
practice. However, the semantics of alm_start is such that an already
started alarm is canceled, thus we need to cancel it in this case, too.